feat: add collection and link management components
- Implemented CollectionCreateFormView for creating new collections.
- Added CollectionLinksManager for managing links within a collection.
- Created CollectionLinksManagerView for displaying and interacting with links.
- Developed DashboardCollectionsManager for managing user collections on the dashboard.
- Introduced DashboardCollectionsManagerView for presenting collections with filtering and editing capabilities.
- Added LinkCreateForm for adding new links to collections.
- Created PublicCatalog for displaying public collections with search and pagination.
- Implemented theme toggle components for user interface customization.
- Added sortable views for collections and links to enhance user experience.
