+++ src/components/theme-toggle/view.tsx
@@ -1,0 +1,46 @@
+"use client";
+
+import { LaptopIcon, MoonIcon, SunIcon } from "lucide-react";
+import { Button } from "@/components/ui/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+
+export interface ThemeToggleViewProps {
+  onSetTheme: (theme: string) => void;
+  variant?: "icon" | "ghost";
+}
+
+export function ThemeToggleView({ onSetTheme, variant = "icon" }: ThemeToggleViewProps) {
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant={variant === "icon" ? "outline" : "ghost"} size="icon" aria-label={variant === "icon" ? "Toggle theme" : "Theme options"}>
+          {variant === "icon" ? (
+            <>
+              <SunIcon className="size-4 scale-100 rotate-0 transition-all dark:scale-0 dark:-rotate-90" />
+              <MoonIcon className="absolute size-4 scale-0 rotate-90 transition-all dark:scale-100 dark:rotate-0" />
+            </>
+          ) : (
+            <LaptopIcon className="size-4" />
+          )}
+          <span className="sr-only">Toggle theme</span>
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent align="end">
+        <DropdownMenuItem onClick={() => onSetTheme("light")}>
+          Light
+        </DropdownMenuItem>
+        <DropdownMenuItem onClick={() => onSetTheme("dark")}>
+          Dark
+        </DropdownMenuItem>
+        <DropdownMenuItem onClick={() => onSetTheme("system")}>
+          System
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
